pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Ivar Carrera <ivar.carrera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
hypercorn = "*"
python-dotenv = "*"
asyncpg = "*"


[tool.poetry.group.test.dependencies]
pytest = "*"
parameterized = "*"
pytest-cov = "*"
httpx = "*"
alt-pytest-asyncio = "*"
pytest-asyncio = "^0.21.0"


[tool.poetry.group.lint.dependencies]
blue = "*"
isort = "*"
ruff = "*"
pip-audit = "*"
mypy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.blue]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = [
    "A",
    "ARG",
    "B",
    "C",
    "C4",
    "E",
    "ERA",
    "F",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = ["A003"]
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"test/**" = ["ARG", "S"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
